[fix] Compilation in Ubuntu 16.04

+++ src/VNLBayes/VideoNLBayes.cpp
@@ -691,8 +691,8 @@ unsigned estimateSimilarPatches(
 			int ct0 = ct[dt - dir];
 			float cx_f = cx0 + (use_flow ? (dir > 0 ? fflow(cx0,cy0,ct0,0) : bflow(cx0,cy0,ct0,0)) : 0.f);
 			float cy_f = cy0 + (use_flow ? (dir > 0 ? fflow(cx0,cy0,ct0,1) : bflow(cx0,cy0,ct0,1)) : 0.f);
-			cx[dt] = std::max(0.f, std::min((float)sz.width  - 1, round(cx_f)));
-			cy[dt] = std::max(0.f, std::min((float)sz.height - 1, round(cy_f)));
+			cx[dt] = std::max(0.f, std::min((float)sz.width  - 1, roundf(cx_f)));
+			cy[dt] = std::max(0.f, std::min((float)sz.height - 1, roundf(cy_f)));
 			ct[dt] = qt;
 		}
 		else
@@ -747,7 +747,7 @@ unsigned estimateSimilarPatches(
 	std::partial_sort(distance.begin(), distance.begin() + nSimP,
 	                  distance.end(), compareFirst);
 
-	if (nSimP <  params.nSimilarPatches)
+	if (nSimP < params.nSimilarPatches)
 		printf("SR2 [%d,%d,%d] ~ nsim = %d\n", px,py,pt,nSimP);
 
 	// Add more patches if their distance is bellow the threshold
